+++ export_academy/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ExportAcademyConfig(AppConfig):
+    name = 'export_academy'